+++ NEWS.md
@@ -1,3 +1,5 @@
+# rphylopic (development version)
+
 # rphylopic 1.1.1
 
 * Minor fixes for Fedora