+++ NEWS.md
@@ -1,5 +1,7 @@
 # rphylopic (development version)
 
+* added preview argument to get_phylopic (#59)
+
 # rphylopic 1.1.1
 
 * Minor fixes for Fedora
@@ -7,7 +9,6 @@
 
 # rphylopic 1.1.0
 
-* added preview argument to get_phylopic (#59)
 * added functions for transforming PhyloPic silhouettes (flipping and rotating)
 * save_phylopic bg argument updated to be "transparent" by default
 * added geom_phylopic (#25)