+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * added text argument to get_attribution (#56)
 * get_attribution now handles multiple uuids
+* added browse_phylopic function (#60)
+* added preview argument to get_phylopic (#59)
 
 # rphylopic 1.1.1