+++ NEWS.md
@@ -6,6 +6,7 @@
 * added preview argument to get_phylopic (#59)
 * switched to {maps} package in base R advanced vignette
 * added filter (license) argument to get_uuid and pick_phylo (#72)
+* added img argument to get_uuid and get_attribution
 
 # rphylopic 1.1.1